refactor(organisms): merge SfStore
