--- packages/vue/src/components/organisms/SfStoreLocator/_internal/SfStore.html
@@ -1,31 +1,0 @@
-<div class="sf-store">
-  <sf-divider class="sf-store__first-divider" />
-  <div class="sf-store__item">
-    <div class="sf-store__item-media" @click="centerOn(latlng)">
-      <sf-image :src="picture" :alt="`${name} picture`" tabindex="0" />
-    </div>
-    <div class="sf-store__item-info">
-      <div class="sf-store__item-info-heading">
-        <div class="sf-store__item-info-heading-name" tabindex="0">
-          {{ name }}
-        </div>
-        <div class="sf-store__item-info-heading-distance" v-if="distance" tabindex="0">
-          <span>{{distance}}km</span> away from you
-        </div>
-      </div>
-      <div class="sf-store__item-info-address" v-if="address" tabindex="0">
-        {{ address }}
-      </div>
-      <div class="sf-store__item-info-phone" v-if="phone">
-        <sf-icon icon="phone" size="13px" color="green-primary" />
-        <span tabindex="0">{{ phone }}</span>
-      </div>
-      <div class="sf-store__item-info-email" v-if="email">
-        <sf-icon icon="mail" size="13px" color="green-primary" />
-        <span tabindex="0">{{ email }}</span>
-      </div>
-      <sf-divider class="sf-store__mobile-divider" />
-    </div>
-  </div>
-  <sf-divider class="sf-store__desktop-divider" />
-</div>
